+++ tests/base/test_dataTransfer.py
@@ -1,11 +1,11 @@
 #!/usr/bin/env python
 
-""" Tests for the data handling module.
+""" Tests for the data transfer module.
 
 Note that these tests are not quite ideal, as they rely on the default implementation
 for the configuration and modules (which can change), but using this implementation
 takes much less time than mocking objects. Note that these tests modify the parameters
-of the ``dataHandling`` module since it is slightly easier here than through a ``config.yaml``
+of the ``dataTransfer`` module since it is slightly easier here than through a ``config.yaml``
 which we would need to rely on being in the proper location.
 
 .. code-author: Raymond Ehlers <raymond.ehlers@yale.edu>, Yale University
@@ -18,41 +18,41 @@ import os
 import logging
 logger = logging.getLogger(__name__)
 
-import overwatch.base.dataHandling as dataHandling
+import overwatch.base.dataTransfer as dataTransfer
 
 @pytest.fixture
-def dataHandlingSetup():
-    """ Basic variables for testing the data handling module.
+def dataTransferSetup():
+    """ Basic variables for testing the data transfer module.
 
     Note:
         We cleanup any possible transferred files stored at ``destination/filename`` after running
         the test.
 
     Note:
-        We modify some of the configuration parameters of the ``dataHandling`` module to simplify testing.
+        We modify some of the configuration parameters of the ``dataTransfer`` module to simplify testing.
         However, this means that those values are permanently changed for that module during testing!
 
     Args:
         None.
     Yields:
-        tuple: (directory, destination) where directory (str) is the path to the data handling
-            test files directory, and destination (str) is the path to the data handling test files
+        tuple: (directory, destination) where directory (str) is the path to the data transfer
+            test files directory, and destination (str) is the path to the data transfer test files
             copy destination directory.
     """
-    directory = os.path.join(os.path.dirname(os.path.realpath(__file__)), "dataHandlingTestFiles")
+    directory = os.path.join(os.path.dirname(os.path.realpath(__file__)), "dataTransferTestFiles")
     destination = os.path.join(directory, "destination")
     # Ensure that it exists. It won't by default because we don't store any files that are copied there in git.
     if not os.path.exists(destination):
         os.makedirs(destination)
-    filenames = dataHandling.determineFilesToMove(directory = directory)
+    filenames = dataTransfer.determineFilesToMove(directory = directory)
 
     # Setup the properties of the module appropriately for testing.
     # debug is to avoid deleting files unintentionally.
-    dataHandling.parameters["debug"] = True
-    dataHandling.parameters["receiverData"] = directory
-    dataHandling.parameters["receiverDataTempStorage"] = os.path.join(directory, "tempStorage")
+    dataTransfer.parameters["debug"] = True
+    dataTransfer.parameters["receiverData"] = directory
+    dataTransfer.parameters["receiverDataTempStorage"] = os.path.join(directory, "tempStorage")
     # Set the site parameters
-    dataHandling.parameters["dataTransferLocations"] = {
+    dataTransfer.parameters["dataTransferLocations"] = {
         "rsync": destination,
         "EOS": destination,
     }
@@ -95,7 +95,7 @@ def testRetryDecorator(loggingMixin, mocker):
     # This attribute is needed for python 2 compatibility. Otherwise, functools will fail when wrapping the object.
     m.__name__ = "mock"
     # Shorter delay to save some time!
-    testFunc = dataHandling.retry(tries = 2, delay = 0.5)(m)
+    testFunc = dataTransfer.retry(tries = 2, delay = 0.5)(m)
     arg = "hello"
     testFunc(arg)
     # Since it was retried twice, it should have been called three times (failed the first time,
@@ -106,22 +106,22 @@ def testRetryDecorator(loggingMixin, mocker):
     # check that the calls occurred with a precise number and signature.
     assert m.mock_calls == [mocker.call(arg), mocker.call(arg), mocker.call(arg)]
 
-def testDetermineFilesToMove(loggingMixin, dataHandlingSetup):
+def testDetermineFilesToMove(loggingMixin, dataTransferSetup):
     """ Test for enumerating the files that should be transferred. """
-    directory, destination, filenames = dataHandlingSetup
+    directory, destination, filenames = dataTransferSetup
 
     assert filenames == ["accepted.root"]
 
 @pytest.mark.parametrize("transferFunction", [
-    dataHandling.copyFilesToOverwatchSites,
-    dataHandling.copyFilesToEOS,
+    dataTransfer.copyFilesToOverwatchSites,
+    dataTransfer.copyFilesToEOS,
 ], ids = ["Overwatch sites", "EOS"])
-def testCopyFilesToOverwatchSites(loggingMixin, dataHandlingSetup, transferFunction):
+def testCopyFilesToOverwatchSites(loggingMixin, dataTransferSetup, transferFunction):
     """ Test for copying files.
 
     For simplicity in testing, we simply copy the files locally.
     """
-    directory, destination, filenames = dataHandlingSetup
+    directory, destination, filenames = dataTransferSetup
     # Remove existing files in the destination to ensure that rsync copies properly.
     # Otherwise, it will not copy any files, which will appear to fail, but only because
     # the destination file already exists.
@@ -144,18 +144,18 @@ def testCopyFilesToOverwatchSites(loggingMixin, dataHandlingSetup, transferFunct
                           filenames = filenames)
 
 @pytest.mark.parametrize("transferFunction", [
-    dataHandling.copyFilesToOverwatchSites,
-    dataHandling.copyFilesToEOS,
+    dataTransfer.copyFilesToOverwatchSites,
+    dataTransfer.copyFilesToEOS,
 ], ids = ["Overwatch sites", "EOS"])
-def testCopyFilesFailure(loggingMixin, dataHandlingSetup, transferFunction, mocker):
+def testCopyFilesFailure(loggingMixin, dataTransferSetup, transferFunction, mocker):
     """ Test failure of copying by providing invalid input files.
 
     This test would be slow because it has to try the copy (but will eventually fail). However, we speed it up by
     replacing ``time.sleep`` with a mock such that it won't sleep during this test.
     """
-    directory, destination, filenames = dataHandlingSetup
+    directory, destination, filenames = dataTransferSetup
     # Speed up the tests by mocking sleep
-    mocker.patch("overwatch.base.dataHandling.time.sleep")
+    mocker.patch("overwatch.base.dataTransfer.time.sleep")
     # We cannot just set an invalid destination because rsync will create that directory.
     filenames = [os.path.join("invalid", f) for f in filenames]
     failedFilenames = transferFunction(directory = directory,
@@ -169,17 +169,17 @@ def testCopyFilesFailure(loggingMixin, dataHandlingSetup, transferFunction, mock
     False,
     True,
 ], ids = ["Debug disabled", "Debug enabled"])
-def testProcessReceivedFiles(loggingMixin, dataHandlingSetup, mocker, debugEnabled):
+def testProcessReceivedFiles(loggingMixin, dataTransferSetup, mocker, debugEnabled):
     """ Test the process received files driver function. """
-    directory, destination, filenames = dataHandlingSetup
+    directory, destination, filenames = dataTransferSetup
     # Also mock delete so we can test deleting
-    dataHandling.parameters["debug"] = debugEnabled
-    mRemove = mocker.patch("overwatch.base.dataHandling.os.remove")
-    (successfullyTransferred, failedFilenames) = dataHandling.processReceivedFiles()
+    dataTransfer.parameters["debug"] = debugEnabled
+    mRemove = mocker.patch("overwatch.base.dataTransfer.os.remove")
+    (successfullyTransferred, failedFilenames) = dataTransfer.processReceivedFiles()
 
     if debugEnabled is False:
         for f in successfullyTransferred:
-            mRemove.assert_any_call(os.path.join(dataHandling.parameters["receiverData"], f))
+            mRemove.assert_any_call(os.path.join(dataTransfer.parameters["receiverData"], f))
 
     # If everything works, all files should be transferred and none should have failed.
     for siteName, failed in iteritems(failedFilenames):
@@ -190,9 +190,9 @@ def testProcessReceivedFiles(loggingMixin, dataHandlingSetup, mocker, debugEnabl
                           destination = destination,
                           filenames = filenames)
 
-def testProcessReceivedFilesFailure(loggingMixin, dataHandlingSetup, mocker):
+def testProcessReceivedFilesFailure(loggingMixin, dataTransferSetup, mocker):
     """ Test the process received files driver function when transferring files fail. """
-    directory, destination, filenames = dataHandlingSetup
+    directory, destination, filenames = dataTransferSetup
     # Mock the transfer functions so this test runs faster and so we can return the filenames as if they
     # failed without running into trouble with our copy method (which would otherwise be unable to find
     # the files to copy).
@@ -200,9 +200,9 @@ def testProcessReceivedFilesFailure(loggingMixin, dataHandlingSetup, mocker):
     #       the copy methods later. And we cannot give a non-existent destination because rsync will just
     #       create the directory. Thus, mocking the initial transfer functions to appear as if they failed
     #       seems to be our best option. Plus, we don't have to wait for the retries, so it's much faster!
-    mocker.patch("overwatch.base.dataHandling.copyFilesToOverwatchSites", return_value = filenames)
-    mocker.patch("overwatch.base.dataHandling.copyFilesToEOS", return_value = filenames)
-    (successfullyTransferred, failedFilenames) = dataHandling.processReceivedFiles()
+    mocker.patch("overwatch.base.dataTransfer.copyFilesToOverwatchSites", return_value = filenames)
+    mocker.patch("overwatch.base.dataTransfer.copyFilesToEOS", return_value = filenames)
+    (successfullyTransferred, failedFilenames) = dataTransfer.processReceivedFiles()
 
     # Everything should fail to transfer
     assert successfullyTransferred == []
@@ -211,6 +211,6 @@ def testProcessReceivedFilesFailure(loggingMixin, dataHandlingSetup, mocker):
     for siteName, failed in iteritems(failedFilenames):
         assert failed == filenames
         checkTransferredFiles(source = directory,
-                              destination = os.path.join(dataHandling.parameters["receiverDataTempStorage"], siteName),
+                              destination = os.path.join(dataTransfer.parameters["receiverDataTempStorage"], siteName),
                               filenames = filenames)